meeting event delete added

+++ controllers/meeting.js
@@ -8,7 +8,11 @@ const {
   _getAvailabilityByBody,
 } = require("./availability");
 const { NOT_FOUND, UNPROCESSABLE_DATA } = require("../constants/errorCodes");
-const { createEvent, createMeeting } = require("../helpers/meeting");
+const {
+  createEvent,
+  createMeeting,
+  deleteEvent,
+} = require("../helpers/meeting");
 const { WorkExperience } = require("../models/workExperience");
 const MIXPANEL_TRACK = require("../helpers/mixpanel");
 
@@ -158,6 +162,9 @@ const cancelMeeting = asyncWrapper(async (req, res) => {
   const { meetingId, userId } = req.body;
   const found = await Meeting.findByPk(meetingId);
   if (!found) res.fail("Meeting data not found");
+  const deletedEvent = await deleteEvent(found.event);
+  console.log("Deleted event: ", deletedEvent);
+  if (!deletedEvent) return res.fail("Meeting could not be deleted");
   await Meeting.destroy({ where: { id: meetingId } });
   if (userId === found.acceptor || userId === found.initiator) {
     const availability = await _getAvailabilityByBody({
